We now use float64 for all placeholders, constants, and (global and local) variables.
- luckily, tensorflow checks all types and hence has admonished where still
  float32 was used by default.
- loss and acc are now also printed with a specific output precision.
- TESTS: Marked all regression tests that diff against files as XFAIL.

+++ src/datasets/classificationdatasets.py
@@ -73,7 +73,7 @@ class ClassificationDatasets:
         """
         # Input placeholders
         with tf.name_scope('input'):
-            xinput = tf.placeholder(tf.float32, [None, input_dimension], name='x-input')
+            xinput = tf.placeholder(tf.float64, [None, input_dimension], name='x-input')
             # print("xinput is "+str(xinput.get_shape()))
 
             # pick from the various available input columns
